requirements.txt
# requirements.txt — Produção

# —— Core Frameworks ——
fastapi
starlette
uvicorn

# —— Autenticação e Criptografia ——
passlib[bcrypt]
cryptography
python-jose       
PYJWT

# —— Validação ——
pydantic==2.11.3
pydantic-settings
email_validator
python-multipart

# —— Banco de Dados ——
sqlalchemy-utils
sqlalchemy
asyncpg
greenlet

# —— Ferramentas Web ——
requests
websockets
python-dotenv

# —— Async e Background ——
anyio
h11
nest-asyncio
sniffio
watchfiles

# —— Visualização e Dashboards ——
dash
dash-bootstrap-components
plotly
fpdf2
XlsxWriter

# —— Dados e DataFrames ——
pandas
numpy
pytz
tzdata
python-dateutil

# —— Utilitários ——
defusedxml
dnspython
pyasn1
rsa
six
retrying
idna
urllib3
charset-normalizer
certifi
colorama
platformdirs
zipp
importlib_metadata
MarkupSafe
Jinja2
itsdangerous
click

# —— Outros ——
fonttools
pillow
narwhals
ecdsa
cffi
pycparser
typer
rich


